Add p5 files_iteration1
This effect is an extension of the group project. It implements an animation effect in which, over time, four buildings and their different levels are gradually revealed. Each building is divided into six levels, and at specific time frames, the different levels of the buildings are progressively shown. "framecount"is used to control the timing, ensuring that each level appears in sequence from the first to the sixth, showcasing a gradual construction process.